feature(framework): Introduce UI5MinimalElement
