+++ packages/base/src/Theming.js
@@ -34,7 +34,7 @@ const _executeThemeChangeCallbacks = theme => {
 const getEffectiveStyle = ElementClass => {
 	const tag = ElementClass.getMetadata().getTag();
 	const customStyle = getCustomCSS(tag) || "";
-	let componentStyles = ElementClass.styles;
+	let componentStyles = ElementClass.styles || "";
 
 	if (Array.isArray(componentStyles)) {
 		componentStyles = componentStyles.join(" ");
