+++ packages/main/src/Avatar.js
@@ -1,4 +1,4 @@
-import UI5Element from "@ui5/webcomponents-base/dist/UI5Element.js";
+import UI5MinimalElement from "@ui5/webcomponents-base/dist/UI5MinimalElement.js";
 import litRender from "@ui5/webcomponents-base/dist/renderer/LitRenderer.js";
 
 // Template
@@ -164,12 +164,12 @@ const metadata = {
  * @constructor
  * @author SAP SE
  * @alias sap.ui.webcomponents.main.Avatar
- * @extends UI5Element
+ * @extends UI5MinimalElement
  * @tagname ui5-avatar
  * @since 1.0.0-rc.6
  * @public
  */
-class Avatar extends UI5Element {
+class Avatar extends UI5MinimalElement {
 	static get metadata() {
 		return metadata;
 	}
