+++ packages/main/src/MessageStrip.js
@@ -1,4 +1,4 @@
-import UI5Element from "@ui5/webcomponents-base/dist/UI5Element.js";
+import UI5MinimalElement from "@ui5/webcomponents-base/dist/UI5MinimalElement.js";
 import litRender from "@ui5/webcomponents-base/dist/renderer/LitRenderer.js";
 import { fetchI18nBundle, getI18nBundle } from "@ui5/webcomponents-base/dist/i18nBundle.js";
 import { isEnter, isSpace } from "@ui5/webcomponents-base/src/events/PseudoEvents.js";
@@ -125,12 +125,12 @@ const metadata = {
  * @constructor
  * @author SAP SE
  * @alias sap.ui.webcomponents.main.MessageStrip
- * @extends UI5Element
+ * @extends UI5MinimalElement
  * @tagname ui5-messagestrip
  * @public
  * @since 0.9.0
  */
-class MessageStrip extends UI5Element {
+class MessageStrip extends UI5MinimalElement {
 	static get metadata() {
 		return metadata;
 	}
