+++ packages/main/src/Tab.js
@@ -1,4 +1,4 @@
-import UI5Element from "@ui5/webcomponents-base/dist/UI5Element.js";
+import UI5MinimalElement from "@ui5/webcomponents-base/dist/UI5MinimalElement.js";
 import litRender from "@ui5/webcomponents-base/dist/renderer/LitRenderer.js";
 import SemanticColor from "./types/SemanticColor.js";
 import Icon from "./Icon.js";
@@ -124,7 +124,7 @@ const metadata = {
  * @tagname ui5-tab
  * @public
  */
-class Tab extends UI5Element {
+class Tab extends UI5MinimalElement {
 	static get metadata() {
 		return metadata;
 	}
