+++ packages/main/src/TabSeparator.js
@@ -1,4 +1,4 @@
-import UI5Element from "@ui5/webcomponents-base/dist/UI5Element.js";
+import UI5MinimalElement from "@ui5/webcomponents-base/dist/UI5MinimalElement.js";
 import litRender from "@ui5/webcomponents-base/dist/renderer/LitRenderer.js";
 import TabSeparatorTemplate from "./generated/templates/TabSeparatorTemplate.lit.js";
 
@@ -20,7 +20,7 @@ const metadata = {
  * @tagname ui5-tab-separator
  * @public
  */
-class TabSeparator extends UI5Element {
+class TabSeparator extends UI5MinimalElement {
 	static get metadata() {
 		return metadata;
 	}
