+++ packages/main/src/TableCell.js
@@ -1,4 +1,4 @@
-import UI5Element from "@ui5/webcomponents-base/dist/UI5Element.js";
+import UI5MinimalElement from "@ui5/webcomponents-base/dist/UI5MinimalElement.js";
 import litRender from "@ui5/webcomponents-base/dist/renderer/LitRenderer.js";
 import TableCellTemplate from "./generated/templates/TableCellTemplate.lit.js";
 
@@ -59,11 +59,11 @@ const metadata = {
  * @constructor
  * @author SAP SE
  * @alias sap.ui.webcomponents.main.TableCell
- * @extends sap.ui.webcomponents.base.UI5Element
+ * @extends sap.ui.webcomponents.base.UI5MinimalElement
  * @tagname ui5-table-cell
  * @public
  */
-class TableCell extends UI5Element {
+class TableCell extends UI5MinimalElement {
 	static get metadata() {
 		return metadata;
 	}
