+++ packages/main/src/TimelineItem.js
@@ -1,4 +1,4 @@
-import UI5Element from "@ui5/webcomponents-base/dist/UI5Element.js";
+import UI5MinimalElement from "@ui5/webcomponents-base/dist/UI5MinimalElement.js";
 import litRender from "@ui5/webcomponents-base/dist/renderer/LitRenderer.js";
 import { getRTL } from "@ui5/webcomponents-base/dist/config/RTL.js";
 import Icon from "./Icon.js";
@@ -115,11 +115,11 @@ const metadata = {
  * @constructor
  * @author SAP SE
  * @alias sap.ui.webcomponents.main.TimelineItem
- * @extends UI5Element
+ * @extends UI5MinimalElement
  * @tagname ui5-timeline
  * @public
  */
-class TimelineItem extends UI5Element {
+class TimelineItem extends UI5MinimalElement {
 	static get metadata() {
 		return metadata;
 	}
