+++ packages/main/src/Token.js
@@ -1,4 +1,4 @@
-import UI5Element from "@ui5/webcomponents-base/dist/UI5Element.js";
+import UI5MinimalElement from "@ui5/webcomponents-base/dist/UI5MinimalElement.js";
 import litRender from "@ui5/webcomponents-base/dist/renderer/LitRenderer.js";
 import { getTheme } from "@ui5/webcomponents-base/dist/config/Theme.js";
 import { getRTL } from "@ui5/webcomponents-base/dist/config/RTL.js";
@@ -100,12 +100,12 @@ const metadata = {
  * @constructor
  * @author SAP SE
  * @alias sap.ui.webcomponents.main.Token
- * @extends UI5Element
+ * @extends UI5MinimalElement
  * @tagname ui5-token
  * @usestextcontent
  * @private
  */
-class Token extends UI5Element {
+class Token extends UI5MinimalElement {
 	static get metadata() {
 		return metadata;
 	}
